readyset-sql: Use `quote_style` field for `CollationName`
This struct (formerly enum) will be used in more places in the future.
Matching on variants that don't really matter to the meaning is
excessively verbose (i.e. we just care about the actual
collation/charset name). Additionally, with this change we can now fully
round trip AST nodes including `CollationName` for testing.

Change-Id: I6a6a696438f7fc9edd3d9a464bad397f6bf80716
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/9999
Reviewed-by: Jason Brown <jason.b@readyset.io>
Tested-by: Buildkite CI