buildkite: Add script to fetch artifacts from prior builds
This script downloads a specified artifact from Buildkite if the
artifact can be found on the current or previous build on the branch.
The artifact is downloaded to the same path under which it is stored in
Buildkite.

The primary use of this script is to retrieve test regression files from
prior build failures so they can be used by later test runs.

Change-Id: I6a6a6964820be95f70fe18c6cb98fbbe2084ee82
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10001
Tested-by: Buildkite CI
Reviewed-by: Michael Zink <michael.z@readyset.io>