replicators: fix MySQL ddl vertical test case.
As part of adding support for ADD KEY, we enhanced the MySQL ddl
vertical test case to include a test for adding a key to a table.

However, the test case was not taking into account that column names
might use reserved words, such as `to`. This resulted in the test
producing wrong queries.

This commit fixes the test case by using backticks to quote the column
name.

Change-Id: I2850916658f59ca1c96a85efd52c87ec4f1174dc
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10003
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>