controller: Fix duplicate cache creation corruption
As previously addressed and worked around in Iad22f58c and I4e5aa25d,
duplicate cache creations could corrupt the state in SqlToMirConverter.

Previously, when we would try to make a duplicate cache, we would modify
some node index values, realize the cache already exists, try to reuse
the previous cache instead, fail to undo the modifications we just made,
and then leave the modified graph state pointing at nodes that don't
exist.

The previous 2 patches made changes in the adapter to first drop an
existing cache before recreating it, but as discovered in this new bug,
3 concurrent cache creation requests can provoke the same kind of graph
metadata corruption previously observed.

This was possible because, for each create cache request in the adapter,
a sequence of actions are taken against the controller (probe, remove,
and create), which can be interleaved with the actions from another
request in the adapter.  In this example with 3 concurrent cache
creation requests, all 3 requests in the adapter first probe the
controller to check if the cache already exists, see that it doesn't,
and then all ask the controller to create the same cache.

Now, in the controller we make a copy of the entire SqlIncorporator
state at the beginning of any cache creation and perform our
modifications on the copy.  If we determine that the cache is not a
duplicate, we update our state with the copy.  If we determine that the
cache is a duplicate, we drop the working copy and retain our original
state.

Fixes: REA-5875
Change-Id: I840778a28cd9838bbcee9d190a434087ff835c33
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10096
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>