readyset-tracing: Refactor tracing setup to make it easier to change
The code was extremely verbose simply to allow a corner case which will
currently never be used in real production deployments (no logging to
file, no statement logging, and no tracing) to avoid a dynamic dispatch.
All 8 other combinations (e.g. logging to file but no tracing) were
repetitively instantiated in the normal way. If benchmarking ever shows
that part of the logging infrastructure to be a meaningful area for
improvement, we could special case that one combination, but it made it
almost impossible to reason about or modify the logging setup. Adding a
fourth variable (see future CL adding file instead of directory path
parameter) would also have been a nightmare.

Change-Id: I6a6a6964378a64773797d7137a68e6506d429fbc
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10067
Reviewed-by: Johnathan Davis <jcd@readyset.io>
Tested-by: Buildkite CI