readyset-dataflow: Add buffer to TopK state
Previously, TopK only kept k records in its state.
This meant that every time a record was removed from the top K
records, an upquery would be issued followed by an in-memory
sort of the entire record set.
Also, on every single call to `on_input`, we would create a heap
of the state, basically sorting the entire state all over again.

This change uses the Auxiliary State to store the state (+ a buffer)
in a sorted vector. Eliminating the need to sort on every single call
and keeping a buffer of size k, reducing the number of upqueries
required.

Change-Id: I68fa30787f3612bc059f18f1c9e2c1da2fc97515
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10022
Reviewed-by: Sidney Cammeresi <sac@readyset.io>
Tested-by: Buildkite CI