readyset-tracing: Allow specifying a single file for log output
The previous only available scheme for logging to a file implied using a
rolling logger managing files in a directory, even if the logfile
rotation cadence was `never`, and did not allow setting a filename.

We add another option, not compatible with the log directory or rotation
cadence options, which only ever logs to a specific file, and allows
setting its full filename including extension.

Change-Id: I6a6a6964f404b62f27f456d7cefe3b8c7d47f465
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10069
Reviewed-by: Johnathan Davis <jcd@readyset.io>
Tested-by: Buildkite CI