aggregations: Factor out accumulating operators
Our dataflow aggregation operator code is confusing to add new
operators to as there's messy separation. `GROUP_CONCAT` is in it's
own module, but the enum variant is buried in the `aggregation`
module. `JSON_OBJECT_AGG` behaves more like the GROUP_CONCAT, but is
located in the `aggregation` code, as well.

This refactor CL follows on the previous CL by:
- introducing a distinct enum for acculumating operations (currently
includes GroupConcat and JsonObjectAgg)
- moves the JsonObjectAgg code into the `accumulator` module
- adds a new `MirInnerNode` variant (`Accumulator`, beside the
existing `Aggregation` and `Extremum` variants)
- cleans up all references
- actually print the function name `GroupConcat`, instead of the
previous highly descriptive name `||`, for debugging and
descriptions. This makes graphviz dumps slightly nicer.

Change-Id: I825da0903834cae8c586b3bd291a61be7aff5027
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10076
Tested-by: Buildkite CI
Reviewed-by: Michael Zink <michael.z@readyset.io>