readyset-tracing: Provide an API for updating the log level
This provides a new `set_log_level` function which allows passing in an
string which will be interpreted as an `EnvFilter` directive, the same
syntax as `LOG_LEVEL`, and reload the filter layer at runtime.

It makes use of a global `OnceLock`, as opposed to returning the closure
to the caller or something like that, mostly because it would be
cumbersome to plumb the closure through to all the places where we might
call this function (i.e. the adapter when processing queries for a SQL
interface, and both the adapter and server HTTP servers). This doesn't
limit us in terms of only allowing one handler to exist, because we are
already limited by the fact that only one `tracing` subscriber is
allowed to exist anyway.

I didn't want to go to the hassle of including and configuring
`tracing_mock` to assert (in the included doctest) that it works as
expected, but you can see it if you add a failing assert at the end of
the test. (`--no-capture`/`--show-output` unfortunately doesn't work for
doctests.)

Addresses: REA-4986

Change-Id: I6a6a696422ca1b08df86e57e2cf9c919d43a71ef
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10081
Tested-by: Buildkite CI
Reviewed-by: Johnathan Davis <jcd@readyset.io>