readyset-dataflow: Fix eviction internal error.
This fixes and internal error in un-query where we will try to evict
from base tables when we have translated the pair of up-queries from
a straddled join.
This also reveals another problem as the eviction key might differ
once we have the pair of up-queries. For that we check the arity of
the eviction key.

Fixes: REA-5774
Change-Id: Ib38dabbc7cb6052e216c7864c0baa677a95f06f4
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10132
Reviewed-by: Sidney Cammeresi <sac@readyset.io>
Tested-by: Buildkite CI