Add table status transitions for SHOW READYSET TABLES
This patch implements the state transitions of table statuses for the
new implementation of SHOW READYSET TABLES.  These updates only update
our state machine, which is still not read by anything.

I tried to choose the most authoritative location to drive each
transition.  It took some time to find the best locations, but we
fortunately don't have any duplication or divergences between MySQL and
PostgreSQL.

There was no single place that kept track of when all index compactions
for a table finish, so this patch adds a RunningCompaction struct that
acts as a guard over all running compactions for a given table.  When
all references to it are dropped, we know that the compactions have
completed and online the table.

Addresses: REA-2463, REA-2907, REA-4863, REA-5751
Change-Id: If432acd3d1b99755ea453a2272f194d7ed525ac0
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10201
Tested-by: Buildkite CI
Reviewed-by: Michael Zink <michael.z@readyset.io>