readyset-dataflow: mirror rows to other partial indices
A node might have multiple partial indices, and subsequently multiple
replay paths passing through it. When a response for an upquery is
processed (replay response), we materialize the rows under the current
replay tag. However, if we receive a request coming from a different
partial index, we will be missing the rows as the state for that index
is not populated.

Ideally we should re-work partial state to have the concept of a
primary key, and other indexes (secondary indexes) that are just
a pointer to the primary key (like we do for base tables).

For now, we mirror the rows from replay and updates to the other
partial indices.

Fixes: REA-5224
Change-Id: I2dd29690ecd13fda6d3f74a01c0f2d1ca9ffabfc
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10153
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>