dataflow-state: Fix compaction progress across multiple indices
If multiple indices were added at once, we would overcount the progress
percentage by a factor equal to the number of indices we were
compacting.  We would also run 1 instance of the compaction watcher per
index we were compacting, which was redundant, since they're all reading
from the same log.

Move the compaction watcher into the RunningCompaction guard so we only
run the watcher once per batch of compactions, and also divide the
progress by the number of indices we're compacting to correct for the
overcounting that used to occur when we'd have multiple indices.

Fixes: REA-5751
Change-Id: Iafa7731dba578023fdffa7a33f818b37aa094e0c
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10205
Reviewed-by: Michael Zink <michael.z@readyset.io>
Tested-by: Buildkite CI