readyset-sql-passes: Make RewriteContext a trait
Because some of the fields of the context are intended to be mutable, we
resort to interior mutability with `RefCell`. This allows us to
partially borrow fields of the context without the benefit of true
partial borrowing we had when it was a struct (i.e. we could borrow one
field mutably and one immutably, which is not possible with a trait,
since the fields are only exposed through the interface methods). This
lets us to continue to borrow the heavy bits of the context while also
allowing us to implement this trait for the adapter in a future CL.

Change-Id: I6a6a69646cc71735a24646bd7a1023473469c46d
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10428
Reviewed-by: Vassili Zarouba <vassili@readyset.io>
Tested-by: Buildkite CI