readyset-adapter: Add Query Sampler
This commit adds a Query Sampler thread.

The sampler thread is responsible for probabilistically sampling
queries that were originally sent to Readyset. The intention is to
pro-actively validate the correctness of the Readyset Server.

The sampler thread will run queries against the upstream database and
compare the normalized results with the results returned by the
Readyset Server.

Queries are passed from adapter to sampler via a bounded channel,
which will drop queries if the channel is full.

The sampler thread has two main queues:
- A queue of queries that have been sent from the adapter. Those are
randomly sampled.
- A retry queue. In case of a mismatch, we add queries to a second
queue to be retried.

Re-try logic is implemented by a new struct that keeps track of some
metadata about the query, such as last upstream result hash, the
number or retry attempts, and the query itself.

The retry queue is a double ended queue, since we need to be able
to add elements in the end of the queue and pop elements in the front.
Elements are added in retry order. The retry is based on a timestamp
calculated from the current time + the retry delay.

The retry queue is bounded, and will drop entries if it is full.

We added the following metrics:
- `readyset_query_sampler_queue_len`: The number of queries in the
sampler queue.
- `readyset_query_sampler_queries_sampled`: The number of queries
sampled.
- `readyset_query_sampler_queries_mismatched`: The number of
queries that were sampled but returned different results.
- `readyset_query_sampler_retry_queue_len`: The number of entries in
the retry queue.
- `readyset_query_sampler_retry_queue_full`: The number of times the
retry queue is full.
- `readyset_query_sampler_max_qps_hit`: The number of times the
sampler rate limiter is hit.
- `readyset_query_sampler_reconnects`: The number of times the
sampler reconnects to the upstream database.

We also added new configuration options:
- `SAMPLER_SAMPLE_RATE`: The probability of sampling a query.
- `SAMPLER_RETRY_DELAY_MS`: The delay between retries.
- `SAMPLER_QUEUE_CAPACITY`: The maximum number of queries in the
sampler queue.
- `SAMPLER_MAX_RETRIES`: The maximum number of retries for a query.
- `SAMPLER_MAX_QUERIES_PER_SECOND`: The maximum number of queries to
sample per second to avoid overwhelming the upstream database.

Note: The current implementation is limited to Text Protocol queries.
A further CL will add support for Binary Protocol queries.

Fixes REA-5328.

Release-Note-Core: Added a query sampler thread to the Readyset
  Adapter. This thread will sample queries and compare the results
  with the results returned by the upstream database and report possible
  mismatches.

Change-Id: I95f8f8b15be977e30998b59f0f266cb685a9e99b
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10231
Reviewed-by: Jason Brown <jason.b@readyset.io>
Tested-by: Buildkite CI