readyset-sql-passes: Add adapter rewrite context parameter
While it is currently empty, this new context will later be the
interface for using `SchemaCatalog` in adapter rewrite passes that we
move over from the server.

It is at this point that we can wire the `SchemaCatalog` through the
adapter to pass it in here. There should be no effect on the rewrites
performed, but bit rot and rebasing in the adapter is a pain, so this
lets us get this part committed sooner.

Change-Id: I1f90467139c75eb04cc4155ae7d675336a6a6964
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/11343
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>