readyset-sql-passes: Include disallow row pass in adapter
This one does not need to be included in the server, but is needed in
the adapter for later unnesting.

Easier than the previous one, as the pass doesn't need any schema info.

You may also note this is out of order, and that's just out of
convenience, since it shouldn't affect the results.

Addresses: REA-5933
Change-Id: Ia7e303cbc389c24483edccf2b26093f06a6a6964
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10763
Reviewed-by: Vassili Zarouba <vassili@readyset.io>
Tested-by: Buildkite CI