replicators: Add opportunistic RDS error message
A common misconfiguration of PostgreSQL RDS is to forget to set
rds.logical_replication = 1 in the instance's parameter group.  If this
setting is left disabled, the replicator experiences a rather obscure
error when attempting to establish a replication connection:

    FATAL: no pg_hba.conf entry for host ...

This error comes marked with the SQL state of E28000 (invalid
authorization specification) and is specific enough that I think it
makes sense to add a little bit of context here.  If it occurs, try to
point the user in the right direction by reminding them to enable
logical replication.

Change-Id: I91a00f6fdabe2a12f06854a44fd7dbe2fe0f76ad
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10229
Tested-by: Buildkite CI
Reviewed-by: Marcelo Altmann <marcelo@readyset.io>