buildkite: Fix bug retrieving artifacts from matrixed steps
If a step is matrixed, there will be a single step key for multiple
jobs.  Since those jobs may all produce artifacts with the same name,
the step key is insufficient to identify jobs from which we should
retrieve artifacts.  This update modifies the query to return job ids
and labels of matching steps, filters the jobs for ones that match the
current label, and uses the job id instead of step name for downloading
the artifacts.

Change-Id: I6a6a6964cd8759cd393589c337e046f051cc324f
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10230
Reviewed-by: Michael Zink <michael.z@readyset.io>
Tested-by: Buildkite CI