treewide: Use correct capitalization of Readyset in strings/READMEs
This patch is a mass replacement of "ReadySet" with "Readyset", which is
the capitalization choice we've landed on.  This is mostly to update any
user-facing messages.  Instances where "ReadySet" is used as a keyword
in a string are left alone.

The cloud directory was ignored for now, so this isn't quite treewide.

Change-Id: I4d0d6a10bbc3d355c765462c65b4debb7dadca10
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10237
Reviewed-by: Ron Hough <ron@readyset.io>
Reviewed-by: Sandaruwan Gunathilake <sandaruwan@readyset.io>
Tested-by: Buildkite CI
Terraform-Plan-Review: Ron Hough <ron@readyset.io>
Reviewed-by: Jason Brown <jason.b@readyset.io>