replicators: Add snapshot query comment for MySQL
This commit adds the ability to add a comment to snapshot queries
for MySQL. The comment is added to the snapshot queries by the
caller and is automatically enclosed within the delimiters /* and */.

This is useful to by-pass tools like Percona Toolkit pt-kill which
automatically kills queries that take longer than a threshold.

Fixes REA-5946.
Closes: 1549

Release-Note-Core: Added the ability to add a comment to snapshot
  queries for MySQL.
Change-Id: I9220e91a5c5a42ffe4b98ede5956d8d389478977
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10242
Tested-by: Buildkite CI
Reviewed-by: Michael Zink <michael.z@readyset.io>