treewide: Deny more lints regarding unsafe
Clippy can enforce that only one unsafe operation is performed within
each unsafe block, and that each unsafe block has a `// SAFETY:` comment
and each unsafe `fn` has a `# Safety` section in its doc comment. For
the latter, we also include private functions in this check, which is
not the default, but caught some problems in our codebase.

There is also a rustc lint which can deny unsafe blocks which don't
actually contain unsafe operations. Unlike the newly enabled clippy
lints, this one wasn't tripped anywhere in our codebase, but may as well
turn it on.

Change-Id: I6a6a696497c461080bbe69b09e25e8acb7ad98bc
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10260
Reviewed-by: Sidney Cammeresi <sac@readyset.io>
Tested-by: Buildkite CI