readyset-server: Fix required argument combination
Previously, we'd error out if --feature-materialization-persistence was
explicitly set without --feature-full-materialization also being set,
which wasn't quite right because --feature-full-materialization is on by
default.

The goal was to prevent an invalid combination of enabling
--feature-materialization-persistence while disabling
--feature-full-materialization.  Now, we move the check from parsing to
a bit later when we examine the values, and we only error if the invalid
combination of values is detected.

Change-Id: Id05aa1187ff662ecc19e0fab5f99df81bee6fa44
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10250
Reviewed-by: Jason Brown <jason.b@readyset.io>
Tested-by: Buildkite CI