readyset-util: Fix some unsafe blocks and SAFETY comments
This will allow us to enable more clippy lints in a future CL.

In this case, the boundaries of what should be `unsafe` was pretty
wrong. Moving them one level up is a bit more verbose, but combined with
the `check-private-items` clippy config allows us to be a lot clearer
about these invariants and who is guaranteeing them.

Change-Id: I6a6a69649705be3e3f2cc2234d97468c5e97ade1
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10255
Tested-by: Buildkite CI
Reviewed-by: Sidney Cammeresi <sac@readyset.io>