readyset: Add pre-snapshot verification interface/options
This patch adds a verify module to the readyset crate, which we now
invoke before proceeding with the typical startup process.

If there is a verification error, the process will log it and exit.  If
the verification passes, startup continues like normal.

The new argument --verify can be used to run only the verification and
then exit.  A successful verification will exit with a zero exit code
and a message to stdout, while a failed verification will exit with a
nonzero exit code and a message to stderr.

The new argument --verify-skip can be passed to skip the verification
and continue with the typical startup process.

The arguments --verify and --verify-skip cannot be specified at the same
time.

There are no real verification checks yet, but a verification failure
can be simulated with the new, hidden --verify-fail argument.

Addresses: REA-5937
Change-Id: I7629bc6d4bd795db51a10703ec1ed31e4cdeaa5b
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10594
Reviewed-by: Michael Zink <michael.z@readyset.io>
Tested-by: Buildkite CI