database-utils: Handle root certificate bundles for PostgreSQL
The interface for TlsConnectionBuilder expects individual certificates.
What we had before works fine if there is 1 certificate in the file, but
let's upgrade our processing to parse however many certs are in the
file, and hand each one to the builder as it expects.

Remove the references to DER-encoded certs.  People use PEM-encoded
certs for MySQL and PostgreSQL.

Fixes: REA-5967
Release-Note-Core: Add support for processing a certificate bundle file
  passed to --ssl-root-cert while running in PostgreSQL mode.
Change-Id: I7cd6eb67dde5b8535bb8d11b81f09373f3b44098
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10445
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>