replicators: Pass replica parameters to MySQL server
When Readyset is registering as a replica in the MySQL
server, it provides an option to provide replica paramters
such as hostname, port, username and password. These are
shown when SHOW REPLICAS is run in the master database.

This commit adds following arguments to Readyset:

```
    --report-host
    --report-port
    --report-user
    --report-password
```

These are passed to the MySQL server with replica
registration.

The arguments are intentionally kept same as the MySQL
arguments for replica registration. The variable names use
the "replica" prefix to indicate that these are replica
parameters.

Fixes: #REA-4690
Closes: #1357

Release-Note-Core: Add `--report-host`, `--report-port`,
  `--report-user` and `--report-password` arguments to
  Readyset. These are used to show the replica parameters
  when SHOW REPLICAS is run in the master database.
Change-Id: I7596fadb056fd39c2701f095cfed19a1a712efd4
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10282
Reviewed-by: Marcelo Altmann <marcelo@readyset.io>
Tested-by: Buildkite CI