builtins: Promote date to datetime in `addtime()`
The msyql function `addtime(expr, expr)` returns the same type as the
first `expr` parameter, according to the docs [0]. However, when that
expr is a `date` data type, mysql silently promotes that to a
`datetime`.  This is reasonable as the entire point of this function is
to "add time", and since `date` inherently has no time, promoting the
type makes sense. Readyset, however, was not previously doing so, and
was just returning a `date`.

[0] https://dev.mysql.com/doc/refman/8.4/en/date-and-time-functions.html \
    #function_addtime

Fixes: REA-5963
Change-Id: Ib851da44d5520760b934771ad78d9e316d89ad9a
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10288
Tested-by: Buildkite CI
Reviewed-by: Marcelo Altmann <marcelo@readyset.io>