readyset-logictest: Don't reset database when targeting remote Readyset
This is mainly to support the setup when running under Antithesis, but
is true in general with the remote readyset mode: even if we may need to
reset the state, we can't do so via the logictest runner when readyset
is still running. Antithesis does so by virtue of only ever running one
test case.

Fixes: REA-5958

Change-Id: I6a6a6964322e474864c9738cda64ea8614b6db6a
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10326
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>