readyset-sql: Fix arbitrary DECIMAL type generation
The previous attempt didn't factor in that we were casting the precision
from a u16 to a u8 when generating DECIMAL, and this could result in a
number less than 255.

A better fix would be to allow u16 (or heck why not as many bits as we
feel like) in DECIMAL, since Postgres supports up to 1000; but we'll
have to do that another time.

Change-Id: I6a6a6964727494cf6bf1bbea734a8f9406e0fb05
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10328
Reviewed-by: Jason Brown <jason.b@readyset.io>
Tested-by: Buildkite CI