readyset-logictest: Check queries aren't proxied
We expect all logictest queries to be executed by readyset even if
there's an upstream to proxy to (i.e. when running with replication or a
remote readyset), but if we fail to create a cache or otherwise fail to
execute the query, we may proxy it. After checking the results,
additionally check that the query was not proxied.

Note that as a workaround for REA-4799, we need to ensure that retries
of a query actually retry migrating the query, which may be marked as
unsupported (and therefore would proxy) in the query status cache, even
after a required table has become available over replication.

In a similar vein, we also add a `drop_prepared` method to our
`DatabaseConnection` wrapper so that we can use manually-prepared
statements, dropping them after each execution. This is because the way
we execute statements with bound parameters through the wrapper results
in a prepared statement being created and cached, which means it will
always be proxied if it was unsupported at the first execution. This
only happens in the mysql branch—the postgres protocol and client
actually allow sending bound parameters without preparing the statement,
but the mysql client library we use must prepare a statement, and caches
it until we explicitly drop it. The new API simply lets us manually
prepare and drop statements, so that retries may see the query is
supported; and now, both postgres and mysql are doing the same thing.

Addresses: REA-4799

Change-Id: I6a6a696436e61315cc2743e2ba95af42f22d4892
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10329
Reviewed-by: Jason Brown <jason.b@readyset.io>
Tested-by: Buildkite CI