readyset-logictest: Get full error backtrace in fuzz mode
We want to take the reason (which is a string representation of the
error) and produce a one-line/truncated failure message for use in
Antithesis; but the verbose formatting of `anyhow::Error` which includes
the backtrace puts the root cause at the end of the list of errors, so
it would be difficult to extract it after the fact. Instead, we format
the error with the root cause on the first line (for use in Antithesis)
followed by the verbose formatting with the backtrace (which will be
logged).

Change-Id: I6a6a69641e126c70683be6e2ccbff40765ed0777
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10331
Reviewed-by: Jason Brown <jason.b@readyset.io>
Tested-by: Buildkite CI