mysql: Parse incoming optimized `TIME` binary format
When sending a `TIME` value over the binary protocol (a prepared
statement), if the value is '00:00:00', the client may send an
optimized value by setting the byte length to 0 and sending no bytes.
See the mysql developer docs [0], section
`ProtocolBinary::MYSQL_TYPE_TIME`.

We weren't handling this optimization case and would fail because we
expected bytes in the buffer, but there were none.

[0] https://dev.mysql.com/doc/dev/mysql-server/latest/
  page_protocol_binary_resultset.html

Fixes: REA-5962
Release-Note-Core: Correctly handle incoming optimized `TIME` binary
  format values when the value is '00:00:00'.
Change-Id: I8c6c7b33de540e6360a0ae5988c06f25e517be4d
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10333
Reviewed-by: Michael Zink <michael.z@readyset.io>
Tested-by: Buildkite CI