readyset-dataflow: Delete sorting of writes in key order
In an attempt to speed processing of writes, we sorted a batch of writes
in key order so that any writes to the same key would be coalesced and
written back only once.

This tactic is all kinds of fail when the key of a row changes, for the
original sequence

    Insert([Int(1), Int(10)])
    Insert([Int(2), Int(20)])
    DeleteByKey { key: [Int(2)] }
    Update { key: [Int(1)], update: [Set(Int(2)), Set(Int(30))] }

is sorted as

    Insert([Int(1), Int(10)])
    Update { key: [Int(1)], update: [Set(Int(2)), Set(Int(30))] }
    Insert([Int(2), Int(20)])
    DeleteByKey { key: [Int(2)] }

which is not equivalent and fails on inserting a duplicate key.

We could try to be more intelligent and not sort when we observe a
sequence like the above (which would take a good bit of new code to
figure out), but since this whole idea is an overoptimization of the
cold path, I think it's better (and certainly safer) to just delete this
code.

Fixes: REA-5125, REA-5127
Change-Id: I53d9ea2602998726b4bc8348a0ea19c176f66444
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10334
Reviewed-by: Michael Zink <michael.z@readyset.io>
Tested-by: Buildkite CI