logictests: Run PostGIS tests with sqlparser
Enabling the PostGIS extension causes several `CREATE VIEW` statements
to be executed, most of which use syntax that is not supported by
nom-sql. This is normally fine, because we can partially parse `CREATE
VIEW` statements and store an error `Result` as the body if it doesn't
succeed. But these statements are *so* unsupported by nom-sql that we
can't even do that (seemingly, we can't recognize a statement separator
after the unparseable bit). Fortunately, sqlparser successfully parses
the statements, and even though we don't support them (and can't even
represent them in our AST) we can store the error `Result` and ignore
them in the replication stream. This way I don't have to fix nom-sql.

Change-Id: I6a6a6964967f88314ac53e76eceec4a9197aa963
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10339
Reviewed-by: Jason Brown <jason.b@readyset.io>
Tested-by: Buildkite CI