replicators: Rework processing of Postgres truncates
Unlike MySQL (in which a truncate is essentially non-transactional and
causes an implicit commit), in Postgres truncates are fully
transactional.  Truncates did not work correctly when they were part of
a larger transaction because the base node expected to get a batch
containing no other writes.

We rewrite the processing of truncates in the Postgres replicator, and
the main part of the code for the main part thereof shrinks in size by
about two thirds after some needed refactoring.

Change-Id: Ice7add2bc26270072a15c5880b8ea590be8cf8ee
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10362
Reviewed-by: Michael Zink <michael.z@readyset.io>
Tested-by: Buildkite CI