readyset-logictest: Always write out test script when fuzzing
If we crash or abort in certain ways, we may not get as far as the
`TestError::Fail` branch which previously would write out the failing
test script. Now, write out the test script before even running it, so
it will be available for debugging.

Also, send it to the `extract_file` handler in abort failures on
Antithesis in case it's useful there too.

Change-Id: I6a6a6964e883da1722daa0b6426db6a281ddba4a
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10368
Reviewed-by: Jason Brown <jason.b@readyset.io>
Tested-by: Buildkite CI