query-generator: Support all BuiltinFunctions
 We were previously only supporting a handful of the variants in
`dataflow_expression::BuiltinFunction`, and we only supported one from
postgres. This CL allows us to support all the builtin function
varaints, and sets us up to support all future entries, as well.

This CL annotates `dataflow_expression::BuiltinFunction` with the
`strum` library, which will createa new enum with simplified versions
of all the  variants. We then use a wrap that in a newtype-like
wrapper struct in `query-generator`, and use dialect-specific function
to filter functions appropriate to each database.

We need to also figure out a dialect-specific version of several
builtin function names as we merge upstream functions with the same
behavior but different names into the same
`dataflow_expression::BuiltinFunction`.

Change-Id: I7e16b38134745b2eecfcf4edfb02f6453c10d514
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10349
Reviewed-by: Michael Zink <michael.z@readyset.io>
Tested-by: Buildkite CI