accumulators: Set collation based on output DfType
For `group_concat()` and `json_object_agg()`, we were hardcoding the
output collation to UTF-8. It's easy enough to figure out the actual
intended collation, we just need to look at the `DfType` that's passed
in to the `over()` function. We still default to UTF-8 lacking
anything else.

Fixes: REA-5948
Release-Note-Core: Use the intended collation for string accumulating
  aggregation function.
Change-Id: Ia2486de5838c162d23cbbd57538faeff90812c0e
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10442
Reviewed-by: Michael Zink <michael.z@readyset.io>
Tested-by: Buildkite CI