query sampler: Sample only queries sent to Readyset
This commits improves the query sampler by only sending queries to the
MPSC channel that have actually been served by readyset. We sample
queries that either succeeded or failed but were not sent as a
fallback to the upstream, as we are interested in finding discrepancies
in the results we sent to clients.

Change-Id: I2697b664198e300624ac054b86f0eaf43c9daf28
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10484
Reviewed-by: Michael Zink <michael.z@readyset.io>
Tested-by: Buildkite CI