post-lookups: Refactor to reuse dataflow functions
The logic internal to accumulating aggregation functions (distinct,
order by) is sufficiently complex that we should reuse it rather
than poorly duplicate it. The CL bends the
PostLookupAggregateFunctions to reuse the dataflow constructs for the
accumulating aggregation functions. The simple aggregation
functions (such as sum, product, min, max) are simple enough to
ignore, however they can be refactored in the future if we choose.

This CL reworks the `readyset-client::view::AggregateIterator` to
reuse the `AccumulationData` struct, which flows neatly into the
`AccumulatorOp` functionality already added to the dataflow
operators. Thus, we reuse the accumulating aggregations across
dataflow and post-lookup aggregates.

Change-Id: Ic315bcc817acdd66e4037c43ce63be787499d8fd
Release-Note-Core: Handle accumulation functions correctly in
  post-lookup aggregation functions.
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10453
Reviewed-by: Michael Zink <michael.z@readyset.io>
Tested-by: Buildkite CI