readyset-adapter: Basic shallow cache result caching and reuse
Because the upstream results are only deciphered by the upper layer of
the protocol, the copy has to happen there.  The copied results are
passed to a guard-like object, which also must be informed of metadata
and be sent a finished signal.  When this object is dropped, if the
correct interactions have been made with it, it inserts its result into
the cache.  Should a bug cause the object to be inadvertently discarded,
no cache modification will occur.

Change-Id: I675306dd92e1b76bdfd9379d0cfefa786933e039
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10474
Reviewed-by: Johnathan Davis <jcd@readyset.io>
Tested-by: Buildkite CI