schema-catalog: Add a synchronizer and handle
The new `SchemaCatalogHandle` will be used in rewrites (REA-5933) and
QueryStatusCache invalidation in the future (REA-6108). For now, it is
simply plumbed through to the `Backend`.

The `SchemaCatalogSynchronizer` is similar to the `ViewSynchronizer` in
implementation only in that it periodically polls the server for the
latest up to date schema information. In the future, we will replace
this with a push model where the server updates any connected adapters,
so don't read too much into the name similarity.

These are introduced together because of the read/write access model:
effectively, only the synchronizer has write access to the underlying
catalog, which it swaps out after getting new versions from the server;
the handle provides only read access for the adapter to use.

Fixes: REA-6110
Addresses: REA-6106
Change-Id: I6a6a6964c481de5c5ef9ae8b657abddb83d80875
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10699
Reviewed-by: Vassili Zarouba <vassili@readyset.io>
Reviewed-by: Jason Brown <jason.b@readyset.io>
Tested-by: Buildkite CI