readyset-adapter: Unwrap optional inner NoriaBackend
This was related to an attempt to allow reconnecting (or retrying
connections?) from the adapter to the server (REA-601) which I dont
think fully materialized. This would not, in my opinion, be the place to
do it anyway. Removing this extra indirection simplifies some code and
makes future work adding more fields to the connector easier.

Addresses: REA-601
Change-Id: I6a6a6964cbde2888e4c6faedf899b052ce9e3b0a
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10496
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>