demo: Fix demo script abort on Ubuntu 22.04
In newer versions of bash, the expression `((retry_count++))` returns an
exit status depending on the result of the arithmetic operation, not
whether the operation was successful.  This can lead to an abort under
certain circumstances, which was happening when running
`readyset_demo.sh` on a devenv instance running Ubuntu 22.04.

Change-Id: I6a6a69640dc8338cd0bce7429736cc2a86bd056c
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10509
Tested-by: Buildkite CI
Reviewed-by: Gautam Gopinadhan <gautam@readyset.io>