readyset-adapter: Differentiate error messages about missing upstream
Most likely, if you see this in a test without an upstream configured,
it's because we tried to run/migrate a query and that failed (i.e. the
`query_fallback` case), but it would be nice if error messages weren't
duplicated and you could be sure which it was without reading 4
different functions.

Change-Id: I6a6a696426561071cf28e7af95db47f489a2052e
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10529
Reviewed-by: Sidney Cammeresi <sac@readyset.io>
Tested-by: Buildkite CI